Azure ML Pipeline/requirements.txt
# pip install -r requirements.txt

# base ----------------------------------------
Cython
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
# opencv-python-headless
Pillow
PyYAML>=5.3
scipy>=1.4.1
tensorboard>=2.2
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0


# logging -------------------------------------
# wandb

# plotting ------------------------------------
seaborn
pandas

# export --------------------------------------
# coremltools==4.0
# onnx>=1.8.0
# scikit-learn==0.19.2  # for coreml quantization

# extras --------------------------------------
thop  # FLOPS computation
pycocotools>=2.0  # COCO mAP

Flask==2.0.1
Flask-RESTful==0.3.9
Flasgger==0.9.5
webargs==8.0.1